requirements.txt
absl-py==2.0.0
accelerate==0.25.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
awscli==1.29.63
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.2
bidict==0.22.1
bleach==6.1.0
blinker==1.7.0
bokeh==3.3.3
botocore==1.31.63
bqplot==0.12.42
build==1.0.3
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.7
colorama==0.4.4
coloredlogs==15.0.1
comm==0.2.0
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
Deprecated==1.2.14
diffusers==0.23.1
diskcache==5.6.3
distro==1.9.0
docutils==0.16
dtaidistance==2.3.11
exceptiongroup==1.1.3
executing==2.0.1
fastapi==0.108.0
filelock==3.12.4
FLAML==2.1.1
Flask==3.0.0
Flask-SocketIO==5.3.6
flatbuffers==23.5.26
fonttools==4.43.1
frozendict==2.4.0
frozenlist==1.4.1
fsspec==2023.9.2
gast==0.5.4
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.62.0
grpcio==1.59.2
h11==0.14.0
h5py==3.10.0
html5lib==1.1
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.20.1
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
ipydatawidgets==4.3.5
ipython==8.17.2
ipyvolume==0.6.3
ipyvue==1.10.1
ipyvuetify==1.8.10
ipywebrtc==0.6.0
ipywidgets==8.1.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jupyterlab-widgets==3.0.9
kaleido==0.2.1
keras==2.14.0
kiwisolver==1.4.5
kubernetes==28.1.0
libclang==16.0.6
lightgbm==4.2.0
linkify-it-py==2.0.2
lxml==5.0.0
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
ml-dtypes==0.2.0
mmh3==4.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
multitasking==0.0.11
networkx==3.1
nltk==3.8.1
npm==0.1.1
numpy==1.26.3
oauthlib==3.2.2
onnxruntime==1.16.3
openai==1.6.1
opencv-python==4.8.1.78
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
opt-einsum==3.3.0
optional-django==0.1.0
overrides==7.4.0
packaging==23.2
pandas==2.1.4
panel==1.3.6
param==2.0.1
parso==0.8.3
peewee==3.17.0
Pillow==10.0.1
posthog==3.1.0
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.7
ptyprocess==0.7.0
pulsar-client==3.4.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.14
pyautogen==0.2.2
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.16.1
pymo==0.2.0
pymongo==3.13.0
PyOpenGL==3.1.7
pyparsing==3.1.1
pypdf==3.17.4
PyPika==0.48.9
pyproject_hooks==1.0.0
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.2
python-multipart==0.0.6
python-socketio==5.11.0
pythreejs==2.4.2
pytz==2023.3.post1
pyviz_comms==3.0.0
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.7.2
s3transfer==0.7.0
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.6
soupsieve==2.5
SpeechRecognition==3.10.1
stack-data==0.6.3
starlette==0.32.0.post1
sympy==1.12
tenacity==8.2.3
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.2.0
tiktoken==0.5.2
tokenizers==0.15.0
tomli==2.0.1
torch==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
traitlets==5.13.0
traittypes==0.2.1
transform3d==0.0.4
transformers==4.36.2
transforms3d==0.4.1
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
uc-micro-py==1.0.2
urllib3==1.26.18
uvicorn==0.25.0
watchfiles==0.21.0
wcwidth==0.2.9
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.0
widgetsnbextension==4.0.9
wrapt==1.14.1
wsproto==1.2.0
xgboost==2.0.3
xyzservices==2023.10.1
yarl==1.9.4
yfinance==0.2.33
zipp==3.17.0
